Refactored mcp server tools into separate files
