+++ unity-mcp-server/src/tools/index.ts
@@ -1,0 +1,14 @@
+import { GetEditorStateTool } from './GetEditorStateTool.js';
+import { ExecuteEditorCommandTool } from './ExecuteEditorCommandTool.js';
+import { GetLogsTool } from './GetLogsTool.js';
+import { Tool } from './types.js';
+
+export * from './types.js';
+
+export function getAllTools(): Tool[] {
+  return [
+    new GetEditorStateTool(),
+    new ExecuteEditorCommandTool(),
+    new GetLogsTool()
+  ];
+}
